requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart>=0.0.12
pydantic>=2.10.0
pydantic-settings>=2.5.0
python-dotenv==1.0.1
networkx==3.3
